Refactoring (#2)
* Refactor
* Enable caching in CI
