+++ include/utils/set_process_name.hpp
@@ -1,0 +1,4 @@
+#pragma once
+#include <string>
+
+void set_process_name(const std::string& name);
